requirements.txt
python-dotenv>=1.0.0
streamlit>=1.32.0
google-generativeai>=0.5.0
sentence-transformers>=2.4.0
faiss-cpu>=1.8.0
PyPDF2>=3.0.0
numpy>=2.0.0